Support for adding new HTTP headers
These changes allow for the inclusion of a new parameter in the
SockJSClient, such that additional HTTP header fields may be defined in
a form of an object. Any existing predefined header fields in code will
be overwritten with the specified HTTP header field of the same header
name (key).  The added headers will be appended onto both the ‘start’
and ‘write’ request variables of XHRStreaming.prototype.